Update test data to convert shorts

+++ src/test/scala/com/databricks/spark/redshift/TestUtils.scala
@@ -55,13 +55,13 @@ object TestUtils {
    */
   val expectedData: Seq[Row] = Seq(
     Row(1.toByte, true, TestUtils.toDate(2015, 6, 1), 1234152.12312498,
-      1.0f, 42, 1239012341823719L, 23, "Unicode's樂趣",
+      1.0f, 42, 1239012341823719L, 23.toShort, "Unicode's樂趣",
       TestUtils.toTimestamp(2015, 6, 1, 0, 0, 0, 1)),
     Row(1.toByte, false, TestUtils.toDate(2015, 6, 2), 0.0, 0.0f, 42,
-      1239012341823719L, -13, "asdf", TestUtils.toTimestamp(2015, 6, 2, 0, 0, 0, 0)),
+      1239012341823719L, -13.toShort, "asdf", TestUtils.toTimestamp(2015, 6, 2, 0, 0, 0, 0)),
     Row(0.toByte, null, TestUtils.toDate(2015, 6, 3), 0.0, -1.0f, 4141214,
       1239012341823719L, null, "f", TestUtils.toTimestamp(2015, 6, 3, 0, 0, 0)),
-    Row(0.toByte, false, null, -1234152.12312498, 100000.0f, null, 1239012341823719L, 24,
+    Row(0.toByte, false, null, -1234152.12312498, 100000.0f, null, 1239012341823719L, 24.toShort,
       "___|_123", null),
     Row(List.fill(10)(null): _*))
   // scalastyle:on
